What's this about:
adding file that needs to be ignored in .gitignore

Adding dependencies in order to make the CLI (View console project) and testing (UnitTesting) works

git pull issue has been resolved, must recreate the project and
re-inserted the directory and the file myself.

Create new class ReviewStatus.cs to adheres to state-based construction
technique.

Create new class DocumentService.cs to handle uncommited (local save) document
progress in an application, modify the Review.cs and DocumentBody.cs so
the DocumentService.cs works, only resolved the null warning issue in
the Document.cs

Fixed the class ReviewUtil.cs, primarily in the processReview
method.

add the solution as things to be ignored in .gitignore